requirements.txt
# Core dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0

# Database
SQLAlchemy>=2.0.0

# Production server (optional)
waitress>=3.0.0

# Testing
pytest>=8.0.0

# Code quality
flake8>=6.1.0
